Move partioning of patient search results to a factory method
This is being done in order support to generate this without needing
an Observable transformer in the search results screen.
