+++ app/src/main/java/org/simple/clinic/searchresultsview/PartitionSearchResultsByVisitedFacility.kt
@@ -5,11 +5,8 @@ import io.reactivex.ObservableSource
 import io.reactivex.ObservableTransformer
 import io.reactivex.rxkotlin.Observables
 import org.simple.clinic.bp.BloodPressureMeasurement
-import org.simple.clinic.bp.PatientToFacilityId
 import org.simple.clinic.facility.Facility
-import org.simple.clinic.patient.FacilityUuid
 import org.simple.clinic.patient.PatientSearchResult
-import org.simple.clinic.patient.PatientUuid
 
 class PartitionSearchResultsByVisitedFacility(
     private val bloodPressureDao: BloodPressureMeasurement.RoomDao,
@@ -24,42 +21,6 @@ class PartitionSearchResultsByVisitedFacility(
         .map(bloodPressureDao::patientToFacilityIds)
 
     return Observables.combineLatest(searchResults, patientToFacilityUuidStream, facilityStream)
-        .map { (patients, patientToFacilities, facility) ->
-          val patientsToVisitedFacilities = mapPatientsToVisitedFacilities(patientToFacilities)
-
-          val (patientsInCurrentFacility, patientsInOtherFacility) = patients.partition { patientSearchResult ->
-            hasPatientVisitedFacility(
-                patientsToVisitedFacilities = patientsToVisitedFacilities,
-                facilityUuid = facility.uuid,
-                patientUuid = patientSearchResult.uuid
-            )
-          }
-
-          PatientSearchResults(
-              visitedCurrentFacility = patientsInCurrentFacility,
-              notVisitedCurrentFacility = patientsInOtherFacility,
-              currentFacility = facility
-          )
-        }
-  }
-
-  private fun hasPatientVisitedFacility(
-      patientsToVisitedFacilities: Map<PatientUuid, Set<FacilityUuid>>,
-      facilityUuid: FacilityUuid,
-      patientUuid: PatientUuid
-  ): Boolean {
-    return patientsToVisitedFacilities[patientUuid]?.contains(facilityUuid) ?: false
-  }
-
-  private fun mapPatientsToVisitedFacilities(patientToFacilities: List<PatientToFacilityId>): Map<PatientUuid, Set<FacilityUuid>> {
-    return patientToFacilities
-        .fold(mutableMapOf<PatientUuid, MutableSet<FacilityUuid>>()) { facilityUuids, (patientUuid, facilityUuid) ->
-          if (patientUuid !in facilityUuids) {
-            facilityUuids[patientUuid] = mutableSetOf()
-          }
-
-          facilityUuids[patientUuid]?.add(facilityUuid)
-          facilityUuids
-        }
+        .map { (patients, patientToFacilities, facility) -> PatientSearchResults.from(patients, patientToFacilities, facility) }
   }
 }
